+++ frontend/apps/web/app/(mgmt)/[account]/new/connection/dynamodb/page.tsx
@@ -1,13 +1,11 @@
 import DynamoDbConnectionForm from '@/components/connections/forms/dynamodb/DynamoDbConnectionForm';
 import OverviewContainer from '@/components/containers/OverviewContainer';
 import PageHeader from '@/components/headers/PageHeader';
-import { useAccount } from '@/components/providers/account-provider';
-import { Connection } from '@neosync/sdk';
-import { useRouter, useSearchParams } from 'next/navigation';
 import { FaAws } from 'react-icons/fa';
+import { useOnCreateSuccess } from '../components/useOnCreateSuccess';
 
 export default async function DynamoDB() {
-  const onSuccess = useOnSuccess();
+  const onSuccess = useOnCreateSuccess();
   return (
     <OverviewContainer
       Header={
@@ -22,39 +20,4 @@ export default async function DynamoDB() {
       <DynamoDbConnectionForm mode="create" onSuccess={onSuccess} />
     </OverviewContainer>
   );
-}
-
-function useOnSuccess(): (conn: Connection) => Promise<void> {
-  const router = useRouter();
-  const { account } = useAccount();
-  // const posthog = usePostHog();
-  const searchParams = useSearchParams();
-  const returnTo = searchParams.get('returnTo');
-
-  return async (conn: Connection): Promise<void> => {
-    if (!account) {
-      return;
-    }
-    if (returnTo) {
-      router.push(returnTo);
-    } else if (conn.id) {
-      router.push(`/${account.name}/connections/${conn.id}`);
-    }
-
-    // try {
-    //   // toast.success('Successfully created OpenAI Connection!');
-    //   // posthog.capture('New Connection Created', { type: 'openai' });
-
-    //   if (returnTo) {
-    //     router.push(returnTo);
-    //   } else if (conn.id) {
-    //     router.push(`/${account.name}/connections/${conn.id}`);
-    //   }
-    // } catch (err) {
-    //   console.error(err);
-    //   toast.error('Unable to create OpenAI Connection', {
-    //     description: getErrorMessage(err),
-    //   });
-    // }
-  };
 }
