--- frontend/apps/web/app/(mgmt)/[account]/new/connection/neon/NeonForm.tsx
@@ -1,6 +1,0 @@
-'use client';
-import PostgresForm from '../postgres/PostgresForm';
-
-export default function NeonForm() {
-  return <PostgresForm />;
-}
