+++ frontend/apps/web/app/(mgmt)/[account]/new/connection/neon/page.tsx
@@ -1,9 +1,13 @@
+'use client';
+import PostgresConnectionForm from '@/components/connections/forms/postgres/PostgresConnectionForm';
 import OverviewContainer from '@/components/containers/OverviewContainer';
 import PageHeader from '@/components/headers/PageHeader';
-import NeonForm from './NeonForm';
+import { ReactElement } from 'react';
+import { useOnCreateSuccess } from '../components/useOnCreateSuccess';
 import { NeonLogo } from './NeonLogo';
 
-export default async function Neon() {
+export default function NewPostgresNeonPage(): ReactElement {
+  const onSuccess = useOnCreateSuccess();
   return (
     <OverviewContainer
       Header={
@@ -15,7 +19,7 @@ export default async function Neon() {
       }
       containerClassName="px-12 md:px-24 lg:px-32"
     >
-      <NeonForm />
+      <PostgresConnectionForm onSuccess={onSuccess} mode="create" />
     </OverviewContainer>
   );
 }
