--- frontend/apps/web/app/(mgmt)/[account]/new/connection/supabase/SupabaseForm.tsx
@@ -1,6 +1,0 @@
-'use client';
-import PostgresForm from '../postgres/PostgresForm';
-
-export default function SupabaseForm() {
-  return <PostgresForm />;
-}
