+++ frontend/apps/web/app/(mgmt)/[account]/new/connection/supabase/page.tsx
@@ -1,9 +1,13 @@
+'use client';
+import PostgresConnectionForm from '@/components/connections/forms/postgres/PostgresConnectionForm';
 import OverviewContainer from '@/components/containers/OverviewContainer';
 import PageHeader from '@/components/headers/PageHeader';
-import SupabaseForm from './SupabaseForm';
+import { ReactElement } from 'react';
+import { useOnCreateSuccess } from '../components/useOnCreateSuccess';
 import { SupabaseLogo } from './SupabaseLogo';
 
-export default async function Supabase() {
+export default function NewSupabasePage(): ReactElement {
+  const onSuccess = useOnCreateSuccess();
   return (
     <OverviewContainer
       Header={
@@ -15,7 +19,7 @@ export default async function Supabase() {
       }
       containerClassName="px-12 md:px-24 lg:px-32"
     >
-      <SupabaseForm />
+      <PostgresConnectionForm onSuccess={onSuccess} mode="create" />
     </OverviewContainer>
   );
 }
